requirements.txt
flask
# gunicorn
sparkpost
python-decouple